Creation of "standard_atm"
"standard_atm.py" returns the geometric and geopotential altitudes and the standard atmosphere temperature, pressure, and density for a provided geometric altitude.  All units are metric, and the input unit is meters (output units are displayed), and the standard atmosphere employed is the 1959 ARDC.  Execution is "standard_atm.py xxx", where "xxx"= geopotential altitude in meters.  Limited testing has shown excellent agreement with the 1959 ARDC, with the only differences noted being a single unit in the final decimal place.  It was run with Python 2.6 and not yet tested for Python 3.0.  Simply replace the #! line with the appropriate pointer to test with a native version of Python.